#2 Вынос dbInitializer в отдельный пакет
