+++ internal/client/app/contract.go
@@ -3,8 +3,6 @@ package app
 import (
 	"context"
 
-	"gorm.io/gorm"
-
 	"github.com/AndrXxX/goph-keeper/internal/client/entities"
 	"github.com/AndrXxX/goph-keeper/pkg/queue"
 )
@@ -17,11 +15,6 @@ type queueRunner interface {
 
 type syncManager interface {
 	Sync(dataType string, updates []any) error
-}
-
-type dbProvider interface {
-	DB(key string) (*gorm.DB, error)
-	RemoveDB() error
 }
 
 type userAccessor interface {
