+++ internal/client/services/dbinitializer/contract.go
@@ -1,0 +1,8 @@
+package dbinitializer
+
+import "gorm.io/gorm"
+
+type dbProvider interface {
+	DB(key string) (*gorm.DB, error)
+	RemoveDB() error
+}
